renamed imager_model package to imaging_model
The package now describes the imaging pipeline, not just the
imager hardware — preparing for Processing and ImagingChain to
land next. Updates imports across opticks/, tests/, notebooks,
and the MTF doc page.
