+++ opticks/imaging_model/aperture.py
