+++ opticks/imaging_model/imager_component.py
