+++ tests/imaging_model/sat_pushbroom_data/binned_test_detector.yaml
