+++ tests/imaging_model/sat_pushbroom_data/block_test_pan_detector.yaml
