+++ tests/imaging_model/sat_pushbroom_data/ms_detector.yaml
