+++ tests/imaging_model/sat_pushbroom_data/optics.yaml
