+++ tests/imaging_model/sat_pushbroom_data/pan_detector.yaml
