+++ tests/imaging_model/sat_pushbroom_data/pan_detector_with_sensor_params.yaml
