+++ tests/imaging_model/sat_pushbroom_data/rw_electronics.yaml
