+++ tests/imaging_model/test_imager.py
@@ -11,13 +11,13 @@ from astropy.units import Unit
 from numpy.testing import assert_allclose
 
 from opticks import u
-from opticks.imager_model.imager import Imager
+from opticks.imaging_model.imager import Imager
 from tests import process_paths
 
 
 class TestImager:
     file_directory = Path("sat_pushbroom_data")
-    alt_file_directory = Path("tests", "imager_model", "sat_pushbroom_data")
+    alt_file_directory = Path("tests", "imaging_model", "sat_pushbroom_data")
     optics_file_path = Path("optics.yaml")
     detector_file_path = Path("pan_detector.yaml")
     rw_electronics_file_path = Path("rw_electronics.yaml")
